# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Avelren API
  version: 0.1.0
  summary: Контракт сервера Avelren / Avelren server contract
  description: |
    UA:
    Read-only API передає Android-клієнту нормалізований стан, отриманий
    сервером із відкритого джерела. Android не звертається до відкритого
    джерела напряму.

    EN:
    This read-only API serves normalized state to the Android client. The
    server obtains data from a publicly accessible external source; Android
    never contacts that source directly.
servers:
  - url: http://localhost:3000
    description: Локальна розробка / Local development
paths:
  /v1/health:
    get:
      operationId: getHealth
      summary: Перевірити стан / Check health
      security: []
      responses:
        "200":
          description: Процес працює / Process is running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                default:
                  externalValue: ./examples/health-response.json
  /v1/workload:
    get:
      operationId: getCurrentWorkload
      summary: Отримати поточний стан / Get current workload
      description: |
        UA: Повертає кешований серверний стан і не запускає опитування відкритого джерела.
        EN: Returns cached server state and never triggers a source poll.
      security: []
      responses:
        "200":
          description: Поточний нормалізований стан / Current normalized state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkloadSnapshot"
              examples:
                default:
                  externalValue: ./examples/current-workload.json
        "503":
          description: Сервер ще не має знімка / Server has no cached workload snapshot yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkloadUnavailable"
              examples:
                no_snapshot:
                  externalValue: ./examples/workload-unavailable.json
  /v1/push/installations:
    post:
      operationId: registerPushInstallation
      summary: Register an Android installation for push delivery
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [installationId, token, platform, locale]
              properties:
                installationId: { type: string, minLength: 22, maxLength: 128 }
                token: { type: string, minLength: 20, maxLength: 4096 }
                platform: { type: string, enum: [android] }
                locale: { type: string, minLength: 2, maxLength: 35 }
      responses:
        '201':
          description: Installation registered; credential is returned only on creation
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [status, installationCredential]
                properties:
                  status: { type: string, enum: [registered] }
                  installationCredential:
                    type: string
                    description: Returned once when a new installation is created
        '400': { description: Invalid request }
        '429': { description: Rate limited }
  /v1/push/installations/{installationId}:
    parameters:
      - $ref: '#/components/parameters/InstallationId'
    patch:
      operationId: updatePushInstallation
      summary: Update installation heartbeat and locale
      security:
        - installationCredential: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [locale]
              properties:
                locale: { type: string, minLength: 2, maxLength: 35 }
      responses:
        '204': { description: Updated }
        '401': { description: Authentication failed }
    delete:
      operationId: disablePushInstallation
      summary: Disable an installation
      security:
        - installationCredential: []
      responses:
        '204': { description: Disabled }
        '401': { description: Authentication failed }
  /v1/push/installations/{installationId}/token:
    parameters:
      - $ref: '#/components/parameters/InstallationId'
    put:
      operationId: rotatePushInstallationToken
      summary: Rotate an installation push token
      security:
        - installationCredential: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token]
              properties:
                token: { type: string, minLength: 20, maxLength: 4096 }
      responses:
        '204': { description: Token rotated }
        '401': { description: Authentication failed }
components:
  securitySchemes:
    installationCredential:
      type: http
      scheme: bearer
      bearerFormat: opaque installation credential
  parameters:
    InstallationId:
      name: installationId
      in: path
      required: true
      schema: { type: string, minLength: 22, maxLength: 128 }
  schemas:
    HealthResponse:
      type: object
      additionalProperties: false
      required:
        - status
        - service
      properties:
        status:
          type: string
          const: ok
        service:
          type: string
          const: avelren-api
    WorkloadSnapshot:
      type: object
      additionalProperties: false
      required:
        - locationId
        - vehicleCount
        - observedAt
        - receivedAt
        - freshness
        - sequence
      properties:
        locationId:
          type: string
          minLength: 1
          maxLength: 128
        vehicleCount:
          type: integer
          minimum: 0
          maximum: 1000000
        observedAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
        freshness:
          type: string
          enum:
            - fresh
            - stale
            - unknown
        sequence:
          type: integer
          minimum: 0
          maximum: 9007199254740991
    WorkloadUnavailable:
      type: object
      additionalProperties: false
      required:
        - error
        - message
        - status
        - timestamp
      properties:
        error:
          type: string
          const: snapshot_unavailable
        message:
          type: string
        status:
          type: integer
          const: 503
        timestamp:
          type: string
          format: date-time